// src/app/profview/profiles.d.ts
export interface ProfileMeta {
      _id: string;
      nc_url: string,
      position_qc: number,
      cycle_number: number,
      dac: string,
      date: Date | moment.Moment | string
      date_added?: Date,
      date_qc?: number,
      max_pres?: number,
      bgcMeasKeys?: string[],
      lat: number,
      lon: number,
      platform_number: number,
      geoLocation?: Point,
      station_parameters: string[],
      station_parameters_in_nc?: string[],
      VERTICAL_SAMPLING_SCHEME?: string,
      PI_NAME?: string,
      WMO_INST_TYPE?: string,
      POSITIONING_SYSTEM?: string,
      DATA_MODE?: string,
      PARAMETER_DATA_MODE?: string[],
      DATA_CENTRE?: string,
      DIRECTION?: string,
      PLATFORM_TYPE?: string,
      pres_max_for_TEMP?: number,
      pres_max_for_PSAL?: number,
      pres_min_for_TEMP?: number,
      pres_min_for_PSAL?: number,
      containsBGC?: boolean,
      isDeep?: boolean,
      BASIN?: number, 
      count?: number,
      core_data_mode?: string,
      roundLat?:string,
      roundLon?:string,
      strLat?:string,
      strLon?:string,
      formatted_station_parameters?: string[]
    }
// ...
export interface StationParameters {
    value: string,
    viewValue: string
  }